+++ src/lib.rs
@@ -20,6 +20,7 @@ pub use element_id::ElementId;
 pub use error::{Error, ErrorPos};
 pub use length::{Length, LengthUnit};
 pub use stream::{TextFrame, Stream};
+pub use tokenize::Tokenize;
 pub use value_id::ValueId;
 pub use values_list::{NumberList, LengthList};
 
@@ -43,11 +44,12 @@ pub mod transform;
 
 mod attribute_id;
 mod attribute_value;
+mod color;
 mod colors;
 mod element_id;
 mod error;
 mod length;
-mod color;
 mod stream;
+mod tokenize;
 mod value_id;
 mod values_list;