+++ src/path.rs
@@ -6,7 +6,7 @@
 //!
 //! [`<path>`]: https://www.w3.org/TR/SVG/paths.html#PathData
 
-use {Stream, TextFrame, Error, ErrorPos};
+use {Tokenize, Stream, TextFrame, Error, ErrorPos};
 
 #[derive(Copy,Clone,Debug,PartialEq)]
 #[allow(missing_docs)]
@@ -88,17 +88,17 @@ pub struct Tokenizer<'a> {
     prev_cmd: Option<u8>,
 }
 
-impl<'a> Tokenizer<'a> {
-    /// Constructs a new `Tokenizer`.
-    pub fn from_frame(frame: TextFrame<'a>) -> Tokenizer<'a> {
+impl<'a> Tokenize<'a> for Tokenizer<'a> {
+    type Token = SegmentToken;
+
+    fn from_frame(frame: TextFrame<'a>) -> Tokenizer<'a> {
         Tokenizer {
             stream: Stream::from_frame(frame),
             prev_cmd: None,
         }
     }
 
-    /// Constructs a new `Tokenizer`.
-    pub fn from_str(text: &'a str) -> Tokenizer<'a> {
+    fn from_str(text: &'a str) -> Tokenizer<'a> {
         Tokenizer {
             stream: Stream::from_str(text),
             prev_cmd: None,
@@ -128,7 +128,7 @@ impl<'a> Tokenizer<'a> {
     ///   into explicit LineTo.
     ///
     ///   Example: `M 10 20 30 40 50 60` -> `M 10 20 L 30 40 L 50 60`
-    pub fn parse_next(&mut self) -> Result<SegmentToken, Error> {
+    fn parse_next(&mut self) -> Result<SegmentToken, Error> {
         let s = &mut self.stream;
 
         s.skip_spaces();